server: move filters into filters package
