--- pkg/server/apiexport_context.go
@@ -1,34 +1,0 @@
-/*
-Copyright 2022 The KCP Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package server
-
-import "context"
-
-type key int
-
-var identityKey key
-
-// WithIdentity adds an APIExport identity to the context.
-func WithIdentity(ctx context.Context, identity string) context.Context {
-	return context.WithValue(ctx, identityKey, identity)
-}
-
-// IdentityFromContext retrieves the APIExport identity from the context, if any.
-func IdentityFromContext(ctx context.Context) string {
-	s, _ := ctx.Value(identityKey).(string)
-	return s
-}
