+++ pkg/server/filters/filters.go
@@ -18,7 +18,6 @@ package filters
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"net/http"
 	"net/url"
@@ -36,9 +35,6 @@ import (
 	kaudit "k8s.io/apiserver/pkg/audit"
 	"k8s.io/apiserver/pkg/endpoints/handlers/responsewriters"
 	"k8s.io/apiserver/pkg/endpoints/request"
-
-	corev1alpha1 "github.com/kcp-dev/kcp/sdk/apis/core/v1alpha1"
-	informersv1alpha1 "github.com/kcp-dev/kcp/sdk/client/informers/externalversions/core/v1alpha1"
 )
 
 type (
@@ -80,49 +76,6 @@ func WithAuditEventClusterAnnotation(handler http.Handler) http.HandlerFunc {
 		cluster := request.ClusterFrom(req.Context())
 		if cluster != nil {
 			kaudit.AddAuditAnnotation(req.Context(), workspaceAnnotation, cluster.Name.String())
-		}
-
-		handler.ServeHTTP(w, req)
-	}
-}
-
-// WithBlockInactiveLogicalClusters ensures that any requests to logical
-// clusters marked inactive are rejected.
-func WithBlockInactiveLogicalClusters(handler http.Handler, kcpClusterClient informersv1alpha1.LogicalClusterClusterInformer) http.HandlerFunc {
-	allowedPathPrefixes := []string{
-		"/openapi",
-		"/apis/core.kcp.io/v1alpha1/logicalclusters",
-	}
-
-	return func(w http.ResponseWriter, req *http.Request) {
-		_, newURL, _, err := ClusterPathFromAndStrip(req)
-		if err != nil {
-			responsewriters.InternalError(w, req, err)
-			return
-		}
-
-		isException := false
-		for _, prefix := range allowedPathPrefixes {
-			if strings.HasPrefix(newURL.String(), prefix) {
-				isException = true
-			}
-		}
-
-		cluster := request.ClusterFrom(req.Context())
-		if cluster != nil && !cluster.Name.Empty() && !isException {
-			logicalCluster, err := kcpClusterClient.Cluster(cluster.Name).Lister().Get(corev1alpha1.LogicalClusterName)
-			if err == nil {
-				if ann, ok := logicalCluster.ObjectMeta.Annotations[inactiveAnnotation]; ok && ann == "true" {
-					responsewriters.ErrorNegotiated(
-						apierrors.NewForbidden(corev1alpha1.Resource("logicalclusters"), cluster.Name.String(), errors.New("logical cluster is marked inactive")),
-						errorCodecs, schema.GroupVersion{}, w, req,
-					)
-					return
-				}
-			} else if !apierrors.IsNotFound(err) {
-				responsewriters.InternalError(w, req, err)
-				return
-			}
 		}
 
 		handler.ServeHTTP(w, req)
